% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_trim.R
\name{adapter_trim}
\alias{adapter_trim}
\title{Trim samples with adapter using cutadapt.}
\usage{
adapter_trim(sample, pair, threads = 4, scRNA = FALSE)
}
\arguments{
\item{sample}{samples with adapter.}

\item{pair}{"single" for single-end (SE) or "paired" for paired-end (PE) reads.}

\item{threads}{the number of threads to be used. Default is 4.}

\item{scRNA}{logic, whether single-cell RNA-seq is quantified or not. Default is FALSE.}
}
\value{
None
}
\description{
Trim samples with adapter using cutadapt based on the results from quality_test().
}
\examples{

adapter_trim(sample = "test_read_2.fastq", pair = "paired", scRNA = TRUE)

}
