% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_free_quan.R
\name{align_free_quan}
\alias{align_free_quan}
\title{Alignment-free expression quantification with Salmon at gene and transcript levels.}
\usage{
align_free_quan(pair, genome, transcript, annotation)
}
\arguments{
\item{pair}{"single" for single-end (SE) or "paired" for paired-end (PE) reads.}

\item{genome}{the reference genome.}

\item{transcript}{the reference transcript}

\item{annotation}{the annotation file.}
}
\value{
None
}
\description{
Alignment-free expression quantification with Salmon at gene and transcript levels.
}
\examples{
\dontrun{

align_free_quan("paired", "sesame.fna", "transcript_sesame.fna","sesame.gff")
}

}
