% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_ge.R
\name{align_ge}
\alias{align_ge}
\title{Aligning the RNA-seq data to the reference genome with HISAT2.}
\usage{
align_ge(pair, taxa, genome, annotation)
}
\arguments{
\item{pair}{"single" for single-end (SE) or "paired" for paired-end (PE) reads.}

\item{taxa}{the scientific or common name of the organism.}

\item{genome}{the reference genome.}

\item{annotation}{the annotation file.}
}
\value{
None
}
\description{
Aligning the RNA-seq data to the reference genome with HISAT2.
}
\examples{
align_ge("paired", "sesame", "sesame.fna", "sesame.gff")
}


