% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dep.R
\name{check_dep}
\alias{check_dep}
\title{Check the existence of dependancies.}
\usage{
check_dep(dependancy)
}
\arguments{
\item{dependancy}{the dependancy to check.}
}
\value{
TRUE if the dependancy exists, otherwise FALSE.
}
\description{
Trim samples with adapter using cutadapt based on the results from quality_test().
}
\examples{

check_dep(dependancy = 'cutadapt')

}
