% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down2quan.R
\name{down2quan}
\alias{down2quan}
\title{Download the RNA-seq samples from the NCBI and produce the quantifications of gene expression.}
\usage{
down2quan(
  accession,
  dir = getwd(),
  taxa,
  novel_transcript = FALSE,
  threads = 4
)
}
\arguments{
\item{accession}{the bioproject accession code for the RNA-seq samples deposited in the NCBI.}

\item{dir}{the working directory. Default is the current working directory.}

\item{taxa}{the scientific or common name of the organism.}

\item{novel_transcript}{logic, whether identifying novel transcripts is expected or not. Default is FALSE.}

\item{threads}{the number of threads to be used. Default is 4.}
}
\value{
None
}
\description{
Download the RNA-seq samples from the NCBI and produce the quantifications of gene expression with alignment-based and alignment-free workflows.
}
\examples{
\dontrun{

down2quan(threads = 4, dir = getwd(), "sesame", novel_transcript = FALSE)
}
}
