% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_Ref.R
\name{down_Ref}
\alias{down_Ref}
\title{Download the latest reference genome, transcript and the corresponding annotation files from NCBI.}
\usage{
down_Ref(taxa)
}
\arguments{
\item{taxa}{the scientific or common name of the organism.}
}
\value{
None
}
\description{
Download the latest reference genome and the corresponding annotation files from NCBI using the NCBI datasets command-line tool.
}
\examples{
down_Ref("sesame")
}
