% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genome.R
\name{extract_genome}
\alias{extract_genome}
\title{Extract the reference genome, transcript and the corresponding annotation files.}
\usage{
extract_genome(taxa)
}
\arguments{
\item{taxa}{the scientific or common name of the organism.}
}
\value{
the extracted reference genome, transcript and annotation file names.
}
\description{
Extract the reference genome, transcript and the corresponding annotation files.
}
\examples{
extract_genome("sesame")
}

