% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_trim.R
\name{get_adapter}
\alias{get_adapter}
\title{Get the adapter based on FASTQC report.}
\usage{
get_adapter(sample)
}
\arguments{
\item{sample}{samples with adapter.}
}
\value{
adapter name
}
\description{
Get the adapter based on FASTQC report produced by quality_test().
}
