% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_trim.R
\name{quality_trim}
\alias{quality_trim}
\title{Trim poor quality samples using cutadapt.}
\usage{
quality_trim(per_base, per_seq, pair, threads = 4)
}
\arguments{
\item{per_base}{samples with poor per base sequence quality.}

\item{per_seq}{samples with poor per sequence quality scores.}

\item{pair}{"single" for single-end (SE) or "paired" for paired-end (PE) reads.}

\item{threads}{the number of threads to be used. Default is 4.}
}
\value{
None
}
\description{
Trim poor quality samples using cutadapt based on the results from quality_test().
}
\examples{
\dontrun{

quality_trim(per_base, per_seq, pair, threads = 4)
}
}
