% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_ass.R
\name{trans_ass}
\alias{trans_ass}
\title{Transcript assembly with StringTie.}
\usage{
trans_ass(novel_transcript = FALSE)
}
\arguments{
\item{novel_transcript}{logic, whether identifying novel transcripts is expected or not. Default is FALSE.}
}
\value{
None
}
\description{
Transcript assembly with StringTie.
}
\examples{
\dontrun{

trans_ass(novel_transcript = FALSE)
}
}
