% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_quan.R
\name{trans_quan}
\alias{trans_quan}
\title{Expression quantification at gene and transcript levels.}
\usage{
trans_quan()
}
\value{
None
}
\description{
Expression quantification at gene and transcript levels.
}
\examples{
\dontrun{

trans_quan()
}
}
