% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{average}
\alias{average}
\title{Take averages with alignment-free and alignment-based quantification
#'}
\usage{
average()
}
\value{
the average of gene expression quantification with alignment-free and alignment-based quantification workflow
}
\description{
Take averages with alignment-free and alignment-based quantification
#'
}
\examples{
average()
}

