% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test.R
\name{qc_test}
\alias{qc_test}
\title{Quality assessments of raw sequence data.}
\usage{
qc_test(threads = 4)
}
\arguments{
\item{threads}{the number of threads to be used. Default is 4.}
}
\value{
Problematic samples' names
}
\description{
Per base sequence quality, per sequence quality scores and adapter content of raw sequence data are assessed and tested based on FastQC.
}
\examples{
\dontrun{

qc_test()
}
}
