% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_quan.R
\name{scRNA_quan}
\alias{scRNA_quan}
\title{scRNA expression quantification with Salmon.}
\usage{
scRNA_quan(transcript, protocol, threads)
}
\arguments{
\item{transcript}{the reference transcript}

\item{protocol}{the single-cell RNA sequencing protocol: dropseq, chromium, or chromiumV3.}
}
\value{
None
}
\description{
scRNA expression quantification with Salmon.
}
\examples{

scRNA_quan(transcript = "test_transcript_Homo_sapiens.fna", protocol = "dropseq")

}
