% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_quan.R
\name{scRNA_quan}
\alias{scRNA_quan}
\title{scRNA expression quantification with Salmon.}
\usage{
scRNA_quan(transcript, protocol)
}
\arguments{
\item{transcript}{the reference transcript}

\item{protocol}{the single-cell RNA sequencing protocol: dropseq, chromium, or chromiumV3.}
}
\value{
None
}
\description{
scRNA expression quantification with Salmon.
}
\examples{
\dontrun{

align_free_quan("paired", "sesame.fna", "transcript_sesame.fna","sesame.gff")
}

}
