% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra_download.R
\name{sra_download}
\alias{sra_download}
\title{Download RNA-seq samples from the NCBI.}
\usage{
sra_download(accession)
}
\arguments{
\item{accession}{the bioproject accession code for the RNA-seq samples deposited in the NCBI.}
}
\description{
Download RNA-seq samples from the NCBI using SRA Toolkit and Entrez Direct based on the accession code.
}
\examples{
sra_download("PRJNA478474")
}
