% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra_download.R
\name{sra_download}
\alias{sra_download}
\title{Download RNA-seq samples from the NCBI.}
\usage{
sra_download(accession, dir = getwd())
}
\arguments{
\item{accession}{the bioproject accession code for the RNA-seq samples deposited in the NCBI.}

\item{dir}{the working directory. Default is the current working directory.}
}
\value{
the status of the download (0 for success, 1 for error).
}
\description{
Download RNA-seq samples from the NCBI using SRA Toolkit and Entrez Direct based on the accession code.
}
\examples{
sra_download(accession = "SRR11427582")

}
